+++ litgpt/data/__init__.py
@@ -4,7 +4,7 @@ from litgpt.data.base import LitDataModule, SFTDataset, get_sft_collate_fn
 from litgpt.data.alpaca import Alpaca
 from litgpt.data.alpaca_2k import Alpaca2k
 from litgpt.data.alpaca_gpt4 import AlpacaGPT4
-from litgpt.data.json import JSON
+from litgpt.data.json_utils import JSON
 from litgpt.data.deita import Deita
 from litgpt.data.dolly import Dolly
 from litgpt.data.flan import FLAN
